move user tests to actions package
