+++ application/gqlgen/gqlgen_test.go
@@ -3,8 +3,6 @@ package gqlgen
 import (
 	"testing"
 
-	"github.com/99designs/gqlgen/client"
-	"github.com/99designs/gqlgen/handler"
 	"github.com/gobuffalo/suite"
 )
 
@@ -21,18 +19,4 @@ func Test_GqlgenSuite(t *testing.T) {
 		Model: model,
 	}
 	suite.Run(t, gs)
-}
-
-func getGqlClient() *client.Client {
-	h := handler.GraphQL(NewExecutableSchema(Config{Resolvers: &Resolver{}}))
-	c := client.New(h)
-	return c
-}
-
-func createFixture(gs *GqlgenSuite, f interface{}) {
-	err := gs.DB.Create(f)
-	if err != nil {
-		gs.T().Errorf("error creating %T fixture, %s", f, err)
-		gs.T().FailNow()
-	}
 }
